o  fixed build controller test
git-svn-id: https://svn.apache.org/repos/asf/maven/continuum/branches/id-refactor@497931 13f79535-47bb-0310-9956-ffa450edef68

+++ continuum-core/src/test/java/org/apache/maven/continuum/buildcontroller/DefaultBuildControllerTest.java
@@ -34,13 +34,13 @@ public class DefaultBuildControllerTest
 {
     private DefaultBuildController controller;
 
-    int projectId1;
+    long projectId1;
 
-    int projectId2;
+    long projectId2;
 
-    int buildDefinitionId1;
+    long buildDefinitionId1;
 
-    int buildDefinitionId2;
+    long buildDefinitionId2;
 
     public void setUp()
         throws Exception
